flask/requirements.txt
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
pycparser==2.20
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.20
uWSGI==2.0.19.1
Werkzeug==1.0.1
